// src/app/api/models/sales-return-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BlankEnum } from '../models/blank-enum';
import { RefundMethodEnum } from '../models/refund-method-enum';
import { ReturnLineRequest } from '../models/return-line-request';

/**
 * Lines are sent whole: on update, a `lines` list replaces what was there.
 * Leave `lines` out to change the header only.
 */
export interface SalesReturnRequest {
  invoice: number;
  lines?: Array<ReturnLineRequest>;
  reason: string;
  refund_method?: (RefundMethodEnum | BlankEnum);
  return_date?: string;
}